Not enough characters to fill the carry buffer
The characters() method in ContentHandlerAdapter is invoked by a SAX
parser when the SAX parser has characters ready for decoding.  The
Base64 decoder requires 4 characters at a time, but the SAX parser can
send any number of characters.  If the number of characters is not a
multiple of 4 then the excess characters are stored in a 'carry buffer'.
The next time the characters() method is called, we move available
characters from the input stream into the carry buffer until there are 4
characters in the carry buffer, then we Base64 decode the carry buffer.
This assumes that there will always be enough characters available to
complete the carry buffer.  In the worst case we might need 3
characters, but the SAX parser has no obligation to send us at least 3
characters.  Sometimes it only sends 1 character.  This causes dcm4che
to throw an ArrayOutOfBoundsException even though the input XML was
valid.

The solution is to copy only as many characters as are available to the
carry buffer, and if that isn't enough to complete the carry buffer then
return and wait for the SAX parser to send some more characters.