IEI-168854 DicomInputStream not cleaning up Inflater
DicomInputStream needs to call end() function on Inflater to clean it up
when close() is being called.
Similarly DicomOutputStream needs to call end() function on Deflater.
This line of code creates an Inflater:
super.in = new InflaterInputStream(super.in, new Inflater(true));

The close() method of InflaterInputStream:
public void close() throws IOException {
    if (!closed) {
        if (usesDefaultInflater)
            inf.end();
        in.close();
        closed = true;
    }
}

So, end() is only called when 'usesDefaultInflater' is true and
'usesDefaultInflater' is only true when the following constructor is
called:
public InflaterInputStream(InputStream in) {
    this(in, new Inflater());
    usesDefaultInflater = true;
}
DicomInputStream is not calling this constructor, so DicomInputStream is
responsible for calling end() on Inflater.
Because end() is not called, end() is being called in the finalizer
which could potentially happen at a much later time. I turns out
Inflater/Deflater allocate memory on the native heap and this memory
does not get released until end() is called. The result is that EI holds
onto large amounts of native heap memory for a long time which could
lead to out of memory situation in native heap.
Here is article which discusses this with potential soution:
https://medium.com/swlh/native-memory-the-silent-jvm-killer-595913cba8e7