Deprecate readDataset methods which accept a len argument
The reason is that the exact behavior is a bit unpredictable as readDataset might also read fileMetaInformation and ignores that when using the len parameter. Also len is an int and not a long.
Instead a LimitedInputStream can be used.

cherry-picked from 3d608c0368906fa1f1bb797b3f734f38726dfc87