Update Transcoder.java
If reading a compressed frame the padding will always be negative and do nothing in the skipFully function because of the negative padding check. 

This is unless the compressed size is actually larger than the uncompressed size (due to bad compression algo selection) and then this will result in a positive number and cause an EOF Exception.

We should only skip if the  frame hasn't been decompressed.